refactoring and added outputgroupboxfunctionality

+++ A01 winform/Models/Circle.cs
@@ -4,14 +4,14 @@ namespace A01_winform.Models
 {
     class Circle : Shape
     {
-        public Circle(string name, int radius) : base(name)
+        public Circle(string name, double radius) : base(name)
         {
             this.Radius = radius;
             UpdateArea();
             SetColor();
         }
 
-        public int Radius { get; set; }
+        public double Radius { get; set; }
 
         private void UpdateArea()
         {
