+++ A01 winform/Models/Square.cs
@@ -4,14 +4,14 @@ namespace A01_winform.Models
 {
     class Square : Shape
     {
-        public Square(string name, int side) : base(name)
+        public Square(string name, double side) : base(name)
         {
             this.Side = side;
             UpdateArea();
             SetColor();
         }
 
-        public int Side { get; set; }
+        public double Side { get; set; }
         private void UpdateArea()
         {
             Area = Math.Pow(Side, 2);
